requirements.txt
openml~=0.12.2
torch~=1.11.0
ConfigSpace~=0.5.0
matplotlib~=3.5.2
numpy~=1.18.2
torchmetrics~=0.9.1
PyYAML~=6.0